feat: enhance backfill process with improved logging, automatic iteration over missing blocks, and new diagnostic script for status checks

+++ src/data/miners/db.py
@@ -1,6 +1,6 @@
 """Database models for miner balance data."""
 
-from sqlalchemy import BigInteger, Column, Index, String
+from sqlalchemy import BigInteger, Column, Index, Numeric, String
 
 from src.helpers.db import Base
 
@@ -12,10 +12,8 @@ class MinerBalanceDB(Base):
 
     block_number = Column(BigInteger, primary_key=True, index=True)
     miner = Column(String(42), nullable=False, index=True)
-    balance_before = Column(BigInteger, nullable=False)  # Wei at block N-1
-    balance_after = Column(BigInteger, nullable=False)  # Wei at block N
-    balance_increase = Column(BigInteger, nullable=False)  # Wei increase
+    balance_before = Column(Numeric, nullable=False)  # Wei at block N-1
+    balance_after = Column(Numeric, nullable=False)  # Wei at block N
+    balance_increase = Column(Numeric, nullable=False)  # Wei increase (can be negative)
 
-    __table_args__ = (
-        Index("idx_miner_block", "miner", "block_number"),
-    )
+    __table_args__ = (Index("idx_miner_block", "miner", "block_number"),)
